+++ ft_memcpy.c
@@ -1,9 +1,9 @@
 #include "libft.h"
 void	*ft_memcpy(void *dest, const void *src, size_t n)
 {
-	unsigned int	i;
-	char			*to;
-	const char		*from;
+	size_t		i;
+	char		*to;
+	const char	*from;
 
 	to = (char *)dest;
 	from = (const char *)src;
