+++ ft_memset.c
@@ -1,8 +1,8 @@
 #include "libft.h"
 void	*ft_memset(void *s, int c, size_t n)
 {
-	unsigned long	i;
-	char			*str;
+	size_t	i;
+	char	*str;
 
 	i = 0;
 	str = (char *)s;
